preparing time

+++ Restaurant/Restaurant/Classes/ApiClasses/MenuSetting.cs
@@ -23,7 +23,7 @@ namespace Restaurant.Classes
         public bool wed { get; set; }
         public bool thur { get; set; }
         public bool fri { get; set; }
-        public Nullable<int> preparingTime { get; set; }
+        public Nullable<decimal> preparingTime { get; set; }
         public Nullable<byte> isActive { get; set; }
         public Nullable<System.DateTime> createDate { get; set; }
         public Nullable<System.DateTime> updateDate { get; set; }
